build: build static binaries with latest pkgs
On nixpkgs bug where windows platform is set incorrectly for static
packages: 281596
This can be worked around by using pkgStatic.pkgsCross rather than
the other way around, or simply by replacing the package.
As it happens, the tls build for windows doesn't need openssl, so
we can remove it entirely and skip this issue altogether.
On darwin, libiconv is failing to build, but also seems to not be
required anymore, and so also removing it!

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>